Here is a working draft. There are two definite things that need to be added:
(1) I need to go over how to do the forward problem in SimPEG. I am floundering around looking for the right functions, but I  am certain this can be understood in 10 minutes or so with some background.

(2) I need to put in the workflow image. I will ask for the best one to use at our next meeting. I assume from I'll use the one from the SimPEG talk, but we can discuss at the next meeting.