Hi Lindsey:
I am having difficulty with the beta cooling method. First, I thought I had a way to do it slightly differently than we discussed, but it does not work. Second, I am trying to implement the algorithm we had on Wednesday, but the "return" function does not work for me. I have tried tab, two, three, and four space indenting, but no dice.