Hey all,
Please note that the directed studies material has been revised.  I have scrapped all previous versions and in order to avoid confusion, I would like to ask you to do the same. At present there are only 3 notebooks worth looking at, and they are the following: "Module 1, Setting up the problem, April", "Module 2, Inversion, April", "Module 3, Choosing Beta, April".

I have tossed the "Module 0" and all mathematical background information. You will notice that certain items were previously to be changed that have not. I will list them here and explain the reasoning:

(1) The workflow image. It was previously suggested that the workflow image be changed, but the one provided has all the points that are covered in the three tutorials.

(2) The python install packages includes a warning. In previous revisions, a method of installing these packages without the warnings was suggested, but for some reason this does not work on my system, so the original order of installation is maintained.

(3) The derivation for obtaining the recovered model is kept as it was. I know that there is a convention aspect at play there, but given that the readership for these tutorials does not extend beyond ourselves, I think that the way it was written, even though not the convention, is not incorrect.

If you plan to view any of the notebooks from here on, please do not bother to read any but the 3 listed above. From now until Tuesday I will go back and look at them, but changes I will make from here will be those dealing with typography and spelling only.

Now that this has been put together, I can see how little ground was covered for the tutorials. My apologies for this. Nevertheless, thank you for having provided me the opportunity to try, however unsuccessful it may have been.